Refactor: Move getRangeOfTextToReplace to utils/textDocument.ts
