+++ server/src/utils/textDocument.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { type Position, type TextDocument } from 'vscode-languageserver-textdocument'
+import { type Position, type Range, type TextDocument } from 'vscode-languageserver-textdocument'
 
 export const getLine = (document: TextDocument, lineNumber: number): string => {
   const range = {
@@ -28,4 +28,35 @@ export const getPreviousCharactersOnLine = (document: TextDocument, position: Po
     end: position
   }
   return document.getText(range)
+}
+
+export const getRangeOfTextToReplace = (
+  document: TextDocument,
+  position: Position,
+  boundRegex: RegExp = /\s|'|"/
+): Range => {
+  const line = getLine(document, position.line)
+
+  const baseIndex = position.character
+
+  let startIndex = baseIndex
+  while (startIndex > 0 && !boundRegex.test(line[startIndex - 1])) {
+    startIndex--
+  }
+
+  let endIndex = baseIndex
+  while (endIndex < line.length && !boundRegex.test(line[endIndex])) {
+    endIndex++
+  }
+
+  return {
+    start: {
+      line: position.line,
+      character: startIndex
+    },
+    end: {
+      line: position.line,
+      character: endIndex
+    }
+  }
 }
